+++ apps/admin-dashboard/app/routes/_dashboard.bull._index.tsx
@@ -7,5 +7,5 @@ export async function loader({ request }: LoaderFunctionArgs) {
     minimumRole: 'owner',
   });
 
-  return redirect('/bull/airtable/jobs');
+  return redirect('/bull/airtable');
 }
