+++ apps/admin-dashboard/app/shared/constants.ts
@@ -8,11 +8,11 @@ const ROUTES = [
   '/applications/:id/accept',
   '/applications/:id/email',
   '/bull',
-  '/bull/:queue/jobs',
+  '/bull/:queue',
   '/bull/:queue/jobs/add',
   '/bull/:queue/jobs/:id',
-  '/bull/:queue/repeatables',
   '/bull/:queue/repeatables/add',
+  '/bull/:queue/repeatables/:id/delete',
   '/events',
   '/events/create',
   '/events/sync-airmeet-event',
